Stripped out Contrib from namespace Minor refactor of project structure Made sure build script generates umbraco packages + nuget packages with dependencies
