+++ Vendr.PaymentProviders.SagePay.sln
@@ -15,10 +15,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Build", "Build", "{A2F6C69B
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "NuSpecs", "NuSpecs", "{81061D69-BD7A-443B-B18D-9BF1F4840A31}"
 	ProjectSection(SolutionItems) = preProject
-		build\NuSpecs\Vendr.Contrib.PaymentProviders.SagePay.nuspec = build\NuSpecs\Vendr.Contrib.PaymentProviders.SagePay.nuspec
+		build\NuSpecs\Vendr.PaymentProviders.SagePay.nuspec = build\NuSpecs\Vendr.PaymentProviders.SagePay.nuspec
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Vendr.Contrib.PaymentProviders.SagePay", "src\Vendr.Contrib.PaymentProviders.SagePay\Vendr.Contrib.PaymentProviders.SagePay.csproj", "{54D75F3D-BD42-45E8-BC99-210A9AD71CEB}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Vendr.PaymentProviders.SagePay", "src\Vendr.PaymentProviders.SagePay\Vendr.PaymentProviders.SagePay.csproj", "{54D75F3D-BD42-45E8-BC99-210A9AD71CEB}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Umbraco", "Umbraco", "{BD551EF3-D31E-4E57-B15D-0E03290C87B8}"
 	ProjectSection(SolutionItems) = preProject
