+++ src/Vendr.PaymentProviders.SagePay/Extensions/ByteArrayExtensions.cs
@@ -1,0 +1,16 @@
+﻿using System.Text;
+
+namespace Vendr.PaymentProviders.SagePay
+{
+    public static class ByteArrayExtensions
+    {
+        public static string ToHex(this byte[] self)
+        {   
+            StringBuilder hex = new StringBuilder(self.Length * 2);
+            foreach (byte b in self)
+                hex.AppendFormat("{0:x2}", b);
+            return hex.ToString();
+            
+        }
+    }
+}
