+++ src/Vendr.PaymentProviders.SagePay/Properties/AssemblyInfo.cs
@@ -5,12 +5,12 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("Vendr.Contrib.PaymentProviders.SagePay")]
-[assembly: AssemblyDescription("")]
+[assembly: AssemblyTitle("Vendr.PaymentProviders.SagePay")]
+[assembly: AssemblyDescription("SagePay payment provider for Vendr, the eCommerce package for Umbrao v8+")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("Vendr.Contrib.PaymentProviders.SagePay")]
-[assembly: AssemblyCopyright("Copyright ©  2019")]
+[assembly: AssemblyCompany("Outfield Digital Ltd")]
+[assembly: AssemblyProduct("Vendr.PaymentProviders.SagePay")]
+[assembly: AssemblyCopyright("Copyright © Outfield Digital Ltd, Spindogs 2020")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
@@ -20,4 +20,4 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("1767d069-22dc-4c19-a293-088679399af4")]
+[assembly: Guid("f334fee4-089f-46e0-90d0-57e94d76902d")]
