+++ src/Vendr.PaymentProviders.SagePay/Properties/VersionInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 
 [assembly: AssemblyVersion("0.1.0.0")]
 [assembly: AssemblyFileVersion("0.1.0.0")]
-[assembly: AssemblyInformationalVersion("0.1.0-alpha.1+Branch.dev.Sha.97539f4e93242f3cbf127e3b9e471e9d932c3847")]
+[assembly: AssemblyInformationalVersion("0.1.0-alpha.5+Branch.dev.Sha.75c98e387a0b73995a9c5045a1caac19787a9605")]
 
 // Generated by the MSBuild WriteCodeFragment class.
