+++ src/Vendr.PaymentProviders.SagePay/Vendr.PaymentProviders.SagePay.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{54D75F3D-BD42-45E8-BC99-210A9AD71CEB}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Vendr.Contrib.PaymentProviders.SagePay</RootNamespace>
-    <AssemblyName>Vendr.Contrib.PaymentProviders.SagePay</AssemblyName>
+    <RootNamespace>Vendr.PaymentProviders.SagePay</RootNamespace>
+    <AssemblyName>Vendr.PaymentProviders.SagePay</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -90,16 +90,17 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Models\CallbackRequestModel.cs" />
+    <Compile Include="Api\Models\CallbackRequestModel.cs" />
+    <Compile Include="Extensions\StringExtensions.cs" />
     <Compile Include="Properties\VersionInfo.cs" />
-    <Compile Include="SagePayClient.cs" />
+    <Compile Include="Api\SagePayClient.cs" />
     <Compile Include="SagePayConstants.cs" />
-    <Compile Include="SagePayInputLoader.cs" />
+    <Compile Include="Helpers\SagePayInputLoader.cs" />
     <Compile Include="SagePayServerPaymentProvider.cs" />
-    <Compile Include="SagePayServerClientConfig.cs" />
+    <Compile Include="Api\Models\SagePayServerClientConfig.cs" />
     <Compile Include="SagePaySettings.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Extensions.cs" />
+    <Compile Include="Extensions\ByteArrayExtensions.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
